+++ src/main/java/com/pischule/mentionbot/model/SentMessage.java
@@ -2,4 +2,4 @@ package com.pischule.mentionbot.model;
 
 import java.time.Instant;
 
-public record SentMessage(long id, Instant createdAt, long chatId, long messageId) {}
+public record SentMessage(long id, Instant createdAt, long chatId, int messageId) {}
